pandaboard:wifi: Enable wifi support for 3.4 kernel
We have added modules to be loaded at boot time.
Some modules correespond to 3.2 and some to 3.4
kernel. Few load modules will fail at boot since
they are not present in that kernel. It won't harm
the boot process but allows us to support 3.2 and
newer kernel in the same branch of this project.

Change-Id: Ib2fffb1dd2d5bfbd9c0aa5e554f0833cccef6a89
Signed-off-by: Vishal Bhoj <vishal.bhoj@linaro.org>