pandaboard: get rid of build variables
get rid of build variables for kernel config and hw graphics.

We were previously setting them based on Build variables.We are
now setting them based on the kernel config available in the
build.Depending on the kernel config available we decide if
kernel supports hardware accelerated graphics. This needs to
be revisited once graphics support is available in 3.4+ kernels.

Change-Id: Ie40ef9e52ebe6a05c8ed89a6059f017ce78b5284
Signed-off-by: Vishal Bhoj <vishal.bhoj@linaro.org>